Move BehaviorDescriptor to polymer dir, finish killing BehaviorDescriptor.tagName
