+++ src/generate-elements.ts
@@ -18,10 +18,11 @@ import * as pathLib from 'path';
 import * as util from 'util';
 
 import {Analysis} from './analysis';
-import {BehaviorDescriptor, Descriptor, DocumentDescriptor, ImportDescriptor, InlineDocumentDescriptor, Property as PropertyDescriptor} from './ast/ast';
+import {Descriptor, DocumentDescriptor, ImportDescriptor, InlineDocumentDescriptor, Property as PropertyDescriptor} from './ast/ast';
 import {Attribute, Element, Elements, Event, Property, SourceLocation} from './elements-format';
 import {JsonDocument} from './json/json-document';
 import {Document} from './parser/document';
+import {BehaviorDescriptor} from './polymer/behavior-descriptor';
 import {PolymerElementDescriptor, PolymerProperty} from './polymer/element-descriptor';
 import {trimLeft} from './utils';
