+++ src/polymer/docs.ts
@@ -15,9 +15,10 @@
 import * as dom5 from 'dom5';
 import * as parse5 from 'parse5';
 
-import {BehaviorDescriptor, BehaviorsByName, Descriptor, EventDescriptor, FeatureDescriptor, Property} from '../ast/ast';
+import {Descriptor, EventDescriptor, FeatureDescriptor, Property} from '../ast/ast';
 import * as jsdoc from '../javascript/jsdoc';
 
+import {BehaviorDescriptor} from './behavior-descriptor';
 import {FunctionDescriptor, PolymerElementDescriptor, PolymerProperty} from './element-descriptor';
